apps/static/Scripts/requirements.txt
arrow==1.2.3
beautifulsoup4==4.8.2
certifi==2023.5.7
chardet==3.0.4
chromedriver==2.24.1
Click==7.0
et-xmlfile==1.1.0
idna==2.8
jdcal==1.4.1
MouseInfo==0.1.3
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.1
Pillow==9.2.0
PyAutoGUI==0.9.53
PyAutoIt==0.6.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2023.3
pywin32==227
pywin32-ctypes==0.2.0
requests==2.22.0
selenium==3.141.0
six==1.16.0
soupsieve==2.4.1
times==0.7
urllib3==1.25.8
